Move NodePairListView into separate class
