+++ src/Neighbor/CMakeLists.txt
@@ -28,6 +28,7 @@ set(Neighbor_headers
     NestedGridNeighborInline.hh
     NodePairIdxType.hh
     NodePairList.hh
+    NodePairListView.hh
     TreeNeighbor.hh
     PairwiseField.hh
     PairwiseFieldElementAccessor.hh
